Package of Activity and Gateways
- Fixed a few style warnings on the way
- Looking for more Javadoc, mine disappeared along the way??
