+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <uses-permission android:name="android.permission.MANAGE_EXTERNAL_STORAGE" />
 
     <application
-        android:name=".MyApp"
+        android:name=".gateway.PlamerContext"
         android:allowBackup="true"
         android:icon="@drawable/icon"
         android:label="@string/app_name"
@@ -15,48 +15,48 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Plamer">
         <activity
-            android:name=".UserPlaylistActivity"
+            android:name=".presenter.UserPlaylistActivity"
             android:exported="false" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".presenter.SearchActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
         <activity
-            android:name=".AddTrackToPlaylistActivity"
+            android:name=".presenter.AddTrackToPlaylistActivity"
             android:exported="false" />
         <activity
-            android:name=".OwnPlaylistActivity"
+            android:name=".presenter.OwnPlaylistActivity"
             android:exported="false" />
         <activity
-            android:name=".UniverseUserActivity"
+            android:name=".presenter.UniverseUserActivity"
             android:exported="false" />
         <activity
-            android:name=".PlaylistActivity"
+            android:name=".presenter.PlaylistActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
         <activity
-            android:name=".TrackActivity"
+            android:name=".presenter.TrackActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingActivity"
+            android:name=".presenter.SettingActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
         <activity
-            android:name=".MainPageActivity"
+            android:name=".presenter.HomepageActivity"
             android:exported="false" />
         <activity
-            android:name=".SignupActivity"
+            android:name=".presenter.SignupActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".presenter.LoginActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
         <activity
-            android:name=".PlayerActivity"
+            android:name=".presenter.PlayerActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".presenter.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
