+++ app/src/main/java/com/mer/plamer/gateway/TinyDB.java
@@ -19,7 +19,7 @@
  *  and unicode 2017 that are used for separating the items in a list.
  */
 
-package com.mer.plamer;
+package com.mer.plamer.gateway;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -206,8 +206,8 @@ public class TinyDB {
      */
     public ArrayList<Integer> getListInt(String key) {
         String[] myList = TextUtils.split(preferences.getString(key, ""), "‚‗‚");
-        ArrayList<String> arrayToList = new ArrayList<String>(Arrays.asList(myList));
-        ArrayList<Integer> newList = new ArrayList<Integer>();
+        ArrayList<String> arrayToList = new ArrayList<>(Arrays.asList(myList));
+        ArrayList<Integer> newList = new ArrayList<>();
 
         for (String item : arrayToList)
             newList.add(Integer.parseInt(item));
@@ -256,8 +256,8 @@ public class TinyDB {
      */
     public ArrayList<Double> getListDouble(String key) {
         String[] myList = TextUtils.split(preferences.getString(key, ""), "‚‗‚");
-        ArrayList<String> arrayToList = new ArrayList<String>(Arrays.asList(myList));
-        ArrayList<Double> newList = new ArrayList<Double>();
+        ArrayList<String> arrayToList = new ArrayList<>(Arrays.asList(myList));
+        ArrayList<Double> newList = new ArrayList<>();
 
         for (String item : arrayToList)
             newList.add(Double.parseDouble(item));
@@ -272,8 +272,8 @@ public class TinyDB {
      */
     public ArrayList<Long> getListLong(String key) {
         String[] myList = TextUtils.split(preferences.getString(key, ""), "‚‗‚");
-        ArrayList<String> arrayToList = new ArrayList<String>(Arrays.asList(myList));
-        ArrayList<Long> newList = new ArrayList<Long>();
+        ArrayList<String> arrayToList = new ArrayList<>(Arrays.asList(myList));
+        ArrayList<Long> newList = new ArrayList<>();
 
         for (String item : arrayToList)
             newList.add(Long.parseLong(item));
@@ -296,7 +296,7 @@ public class TinyDB {
      * @return ArrayList of String
      */
     public ArrayList<String> getListString(String key) {
-        return new ArrayList<String>(Arrays.asList(TextUtils.split(preferences.getString(key, ""), "‚‗‚")));
+        return new ArrayList<>(Arrays.asList(TextUtils.split(preferences.getString(key, ""), "‚‗‚")));
     }
 
     /**
@@ -315,7 +315,7 @@ public class TinyDB {
      */
     public ArrayList<Boolean> getListBoolean(String key) {
         ArrayList<String> myList = getListString(key);
-        ArrayList<Boolean> newList = new ArrayList<Boolean>();
+        ArrayList<Boolean> newList = new ArrayList<>();
 
         for (String item : myList) {
             if (item.equals("true")) {
@@ -333,7 +333,7 @@ public class TinyDB {
     	Gson gson = new Gson();
 
     	ArrayList<String> objStrings = getListString(key);
-    	ArrayList<Object> objects =  new ArrayList<Object>();
+    	ArrayList<Object> objects = new ArrayList<>();
 
     	for(String jObjString : objStrings){
     		Object value  = gson.fromJson(jObjString,  mClass);
@@ -467,7 +467,7 @@ public class TinyDB {
      */
     public void putListBoolean(String key, ArrayList<Boolean> boolList) {
         checkForNullKey(key);
-        ArrayList<String> newList = new ArrayList<String>();
+        ArrayList<String> newList = new ArrayList<>();
 
         for (Boolean item : boolList) {
             if (item) {
@@ -494,7 +494,7 @@ public class TinyDB {
     public void putListObject(String key, ArrayList<Object> objArray){
     	checkForNullKey(key);
     	Gson gson = new Gson();
-    	ArrayList<String> objStrings = new ArrayList<String>();
+    	ArrayList<String> objStrings = new ArrayList<>();
     	for(Object obj : objArray){
     		objStrings.add(gson.toJson(obj));
     	}
