+++ app/src/main/java/com/mer/plamer/presenter/HomepageActivity.java
@@ -1,4 +1,4 @@
-package com.mer.plamer;
+package com.mer.plamer.presenter;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,13 +7,14 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.mer.plamer.R;
 import com.mer.plamer.controller.PlayControl;
 import com.mer.plamer.usecases.PlayAction;
 
 /**
  * Activity to construct the home page view
  */
-public class MainPageActivity extends AppCompatActivity {
+public class HomepageActivity extends AppCompatActivity {
 
     /**
      * Construct view and define actions for each interactive elements
@@ -35,42 +36,42 @@ public class MainPageActivity extends AppCompatActivity {
         // open the setting page
         ImageButton setting = findViewById(R.id.main_settings);
         setting.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this, SettingActivity.class);
+            Intent intent = new Intent(HomepageActivity.this, SettingActivity.class);
             startActivity(intent);
         });
 
         // open the track library
         ImageButton track = findViewById(R.id.main_track_list);
         track.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this, TrackActivity.class);
+            Intent intent = new Intent(HomepageActivity.this, TrackActivity.class);
             startActivity(intent);
         });
 
         // open the playlist
         ImageButton playList = findViewById(R.id.main_playlist);
         playList.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this, PlaylistActivity.class);
+            Intent intent = new Intent(HomepageActivity.this, PlaylistActivity.class);
             startActivity(intent);
         });
 
         // open the search page
         ImageButton search = findViewById(R.id.main_search);
         search.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this, SearchActivity.class);
+            Intent intent = new Intent(HomepageActivity.this, SearchActivity.class);
             startActivity(intent);
         });
 
         // open the user library
         ImageButton users = findViewById(R.id.main_users);
         users.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this,
+            Intent intent = new Intent(HomepageActivity.this,
                     UniverseUserActivity.class);
             startActivity(intent);
         });
 
         // open the playing page
         playing.setOnClickListener(v -> {
-            Intent intent = new Intent(MainPageActivity.this, PlayerActivity.class);
+            Intent intent = new Intent(HomepageActivity.this, PlayerActivity.class);
             startActivity(intent);
         });
 
@@ -86,7 +87,7 @@ public class MainPageActivity extends AppCompatActivity {
 
         // change the loop style
         repeatButton.setOnClickListener(v -> {
-            Toast.makeText(MainPageActivity.this,
+            Toast.makeText(HomepageActivity.this,
                     PlayControl.changePlayMode(), Toast.LENGTH_SHORT).show();
             if (PlayAction.order == PlayAction.PlayOrder.LIST) {
                 ((ImageButton) v).setImageResource(R.drawable.repeat_list);
