+++ app/src/main/java/com/mer/plamer/presenter/SignupActivity.java
@@ -1,4 +1,4 @@
-package com.mer.plamer;
+package com.mer.plamer.presenter;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,7 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.mer.plamer.R;
 import com.mer.plamer.controller.UserControl;
 
 /**
