+++ app/src/main/res/layout/add_track_to_playlist_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddTrackToPlaylistActivity"
+    tools:context=".presenter.AddTrackToPlaylistActivity"
     android:background="@drawable/background">
 
     <ImageButton
