+++ app/src/main/res/layout/search_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchActivity"
+    tools:context=".presenter.SearchActivity"
     android:background="@drawable/background">
 
     <ListView
