+++ app/src/main/res/layout/setting_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SettingActivity"
+    tools:context=".presenter.SettingActivity"
     android:background="@drawable/background">
 
     <ImageButton
