+++ app/src/test/java/com/mer/plamer/controllerTest/TrackLibraryControlTest.java
@@ -1,16 +1,12 @@
 package com.mer.plamer.controllerTest;
 
-import static org.junit.Assert.*;
 import static org.mockito.Mockito.mock;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
-import com.mer.plamer.TinyDB;
+import com.mer.plamer.gateway.TinyDB;
 import com.mer.plamer.controller.TrackLibraryControl;
 import android.media.MediaMetadataRetriever;
-
-import android.media.MediaPlayer;
 
 @RunWith(MockitoJUnitRunner.class)
 public class TrackLibraryControlTest {
