+++ app/src/test/java/com/mer/plamer/controllerTest/UserControlTest.java
@@ -3,18 +3,11 @@ package com.mer.plamer.controllerTest;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.mock;
 
-import com.mer.plamer.TinyDB;
+import com.mer.plamer.gateway.TinyDB;
 import com.mer.plamer.controller.UserControl;
-import com.mer.plamer.usecases.PlaylistLibraryAction;
-import com.mer.plamer.usecases.UserAction;
-import com.mer.plamer.usecases.UserLibraryAction;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.mock.*;
-import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -42,13 +35,6 @@ public class UserControlTest {
         UserControl uc = new UserControl(tinyDB);
         assertEquals("No user founded.", uc.getAccountInfo());
         uc.scanLocal();
-
-    }
-
-    @Test
-    public void testAdminUC() {
-        TinyDB tinyDB = mock(TinyDB.class);
-        UserControl uc = new UserControl(tinyDB);
 
     }
 }
