+++ app/src/test/java/com/mer/plamer/entitiesTest/TrackLibraryTest.java
@@ -4,7 +4,6 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
 
 import com.mer.plamer.entities.Track;
 import com.mer.plamer.entities.TrackLibrary;
