+++ app/src/test/java/com/mer/plamer/presenterTest/TrackAdapterTest.java
@@ -1,4 +1,4 @@
-package com.mer.plamer.controllerTest;
+package com.mer.plamer.presenterTest;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.mock;
@@ -7,9 +7,9 @@ import android.content.Context;
 import android.media.MediaPlayer;
 import android.view.LayoutInflater;
 
-import com.mer.plamer.controller.AddAdapter;
+
 import com.mer.plamer.controller.PlayControl;
-import com.mer.plamer.controller.PlaylistAdapter;
+import com.mer.plamer.presenter.TrackAdapter;
 import com.mer.plamer.entities.Playlist;
 import com.mer.plamer.entities.PlaylistLibrary;
 import com.mer.plamer.entities.Track;
@@ -26,25 +26,24 @@ import org.mockito.runners.MockitoJUnitRunner;
 import java.util.ArrayList;
 
 @RunWith(MockitoJUnitRunner.class)
-public class AddAdapterTest {
-
-    AddAdapter aa;
+public class TrackAdapterTest {
+    TrackAdapter ta;
 
     @Before
     public void setUp() {
         Context ctx = mock(Context.class);
         LayoutInflater lif = mock(LayoutInflater.class);
-        ArrayList<AddAdapter.AddDataHolder> testlist = new ArrayList<>();
-        AddAdapter.AddDataHolder ah = mock(AddAdapter.AddDataHolder.class);
-        testlist.add(ah);
-        aa = new AddAdapter(ctx, testlist, lif);
+        ArrayList<TrackAdapter.TrackDataHolder> talist = new ArrayList<>();
+        TrackAdapter.TrackDataHolder th = mock(TrackAdapter.TrackDataHolder.class);
+        talist.add(th);
+        ta = new TrackAdapter(ctx, lif, talist);
     }
 
     @Test
     public void testProperties() {
-        assertEquals(1, aa.getCount());
-        assertNotNull(aa.getItem(0));
-        assertEquals(0, aa.getItemId(0));
+        assertEquals(1, ta.getCount());
+        assertNotNull(ta.getItem(0));
+        assertEquals(0, ta.getItemId(0));
 
         MediaPlayer mp = mock(MediaPlayer.class);
         PlayAction.setMediaPlayer(mp);
@@ -57,7 +56,6 @@ public class AddAdapterTest {
         t1.setTitle("God");
         t1.setLength("100");
         tl.add(t1);
-        String t1id = t1.getID();
         TrackLibraryAction.assignLibrary(tl);
         PlayAction.setPosition(0);
         PlaylistLibrary pll = new PlaylistLibrary();
@@ -68,6 +66,7 @@ public class AddAdapterTest {
         pll.add(pl2);
         PlaylistLibraryAction.assignLibrary(pll);
 
-        AddAdapter.AddDataHolder aah = new AddAdapter.AddDataHolder(t1id, true);
+
     }
+
 }
